[refactor]: test 부분 제외 PR피드백 수정
