+++ src/main/java/org/c4marathon/assignment/controller/GlobalExceptionHandler.java
@@ -1,0 +1,22 @@
+package org.c4marathon.assignment.controller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class GlobalExceptionHandler {
+
+	// IllegalArgumentException 처리
+	@ExceptionHandler(IllegalArgumentException.class)
+	public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException e) {
+		return ResponseEntity.badRequest().body("잘못된 요청: " + e.getMessage());
+	}
+
+	// 모든 예외 처리
+	@ExceptionHandler(Exception.class)
+	public ResponseEntity<String> handleGeneralException(Exception e) {
+		return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("서버 에러: " + e.getMessage());
+	}
+}
