+++ src/main/java/org/c4marathon/assignment/domain/User.java
@@ -44,15 +44,6 @@ public class User {
 		this.savingAccounts = new ArrayList<>();
 	}
 
-	// 테스트용 생성자: userId 포함!
-	public User(Long userId, String password, String name, String registrationNum) {
-		this.userId = userId;
-		this.password = password;
-		this.name = name;
-		this.registrationNum = registrationNum;
-		this.savingAccounts = new ArrayList<>();
-	}
-
 	//메인 계좌 설정 메서드
 	public void setMainAccount(Account mainAccount) {
 		this.mainAccount = mainAccount;
